tests/requirements.txt
pytest
requests
python-dotenv
groq
rich